Better support for Sprite CGs
Sprite CGs are complex. This commit allows tracking these kinds of CGs
by a identifier and their state. Because changes happen all the time,
changes of state automatically unlock that state in the gallery.

Gallery images that are sprite CGs can be declared as patches to an
initial state. This avoids a lot of repetition, although declaring
gallery items remains quite a lot of work.