Use showif to switch between displayables in the screen.
Ren'Py seems to be having some problem focusing unrelated displayables
when quickly replacing a displayable by another.

Current hypothesis is that when replacing the imagebutton with a `null`
displayable (which isn't focusable), the next focusable displayable gets
the focus, and quickly gets unfocused (because the mouse pointer is
elsewhere). Need to do more tests though.