Reset transitions at the start of an interaction.
This requires Ren'Py 6.99.8 to function.

It fixes an issue where a restarted interaction (which can occur from the CTC indicator being shown, various screed language hovers, etc) could cause the transition to be removed. Ren'Py 6.99.8 now has a function that can detect this condition, making it possible to program around it.